Additional Typo/Word-Choice/Formatting Corrections
I tried not to duplicate any of the items in the other PRs. There are several formatting corrections based on what I perceived to be the document's style choices. I admit to having made at least a few changes based on style preference.

There is a broad consistency issue with the way that HTTP status codes are referenced in that several different styles are used. Sometimes it is only the code itself, sometimes the code plus "HTTP status", sometimes "a status of ___", sometimes "returns a ___" and sometimes "returns ___". I made some initial changes to homogenize these, but eventually stopped as it varied so much.